feat: inbox functionality at the header undread, reply, replyall, start, forward
