+++ client/src/hooks/email/index.ts
@@ -5,3 +5,4 @@ export * from './useToggleFavoriate'
 export * from './useTrashMutate'
 export * from './useArchiveMutate'
 export * from './useDeleteMutate'
+export * from './useThreadAction'
