+++ client/src/hooks/email/useDeleteMutate/useDeleteMutate.types.ts
@@ -1,0 +1,5 @@
+import { IEmail } from 'gmail-api-parse-message-ts'
+
+export type UseDeleteMutateType = {
+  threads: IEmail[]
+}
