+++ client/src/hooks/email/useThreadAction/index.ts
@@ -1,0 +1,2 @@
+export * from './useThreadAction'
+export * from './useThreadAction.types'
