+++ client/src/hooks/email/useThreadAction/useThreadAction.types.ts
@@ -1,0 +1,5 @@
+import { IEmail } from 'gmail-api-parse-message-ts'
+
+export type UseThreadActionType = {
+  items: IEmail[]
+}
